requirements.txt
crewai
crewai_tools
python-decouple
transformers
sentence-transformers
torch
langchain-community
langchain-groq